[backend] extract engine hits converter
