+++ opencti-platform/opencti-graphql/src/database/engine.js
@@ -56,8 +56,6 @@ import {
   ID_STANDARD,
   IDS_STIX,
   isAbstract,
-  REL_INDEX_PREFIX,
-  RULE_PREFIX
 } from '../schema/general';
 import { isModifiedObject, isUpdatedAtObject, } from '../schema/fieldDataAdapter';
 import { getParentTypes, keepMostRestrictiveTypes } from '../schema/schemaUtils';
@@ -73,14 +71,12 @@ import { isBasicRelationship } from '../schema/stixRelationship';
 import { RELATION_INDICATES, RELATION_PUBLISHES, RELATION_RELATED_TO } from '../schema/stixCoreRelationship';
 import { generateInternalId, INTERNAL_FROM_FIELD, INTERNAL_TO_FIELD } from '../schema/identifier';
 import { controlUserRestrictDeleteAgainstElement, executionContext, SYSTEM_USER, userFilterStoreElements } from '../utils/access';
-import { isSingleRelationsRef, } from '../schema/stixEmbeddedRelationship';
 import { now, runtimeFieldObservableValueScript } from '../utils/format';
 import { ENTITY_TYPE_KILL_CHAIN_PHASE, ENTITY_TYPE_MARKING_DEFINITION } from '../schema/stixMetaObject';
 import { getEntitiesListFromCache } from './cache';
 import { ENTITY_TYPE_MIGRATION_STATUS, ENTITY_TYPE_STATUS, ENTITY_TYPE_USER } from '../schema/internalObject';
 import { meterManager, telemetry } from '../config/tracing';
 import { isBooleanAttribute, isDateAttribute, isNumericAttribute, isObjectAttribute, schemaAttributesDefinition, validateDataBeforeIndexing } from '../schema/schema-attributes';
-import { convertTypeToStixType } from './stix-converter';
 import { extractEntityRepresentativeName, extractRepresentative } from './entity-representative';
 import { ENTITY_TYPE_IDENTITY_ORGANIZATION } from '../modules/organization/organization-types';
 import { checkAndConvertFilters, isFilterGroupNotEmpty } from '../utils/filtering/filtering-utils';
@@ -109,6 +105,7 @@ import { FilterMode } from '../generated/graphql';
 import { buildElasticSortingForAttributeCriteria } from '../utils/sorting';
 import { ENTITY_TYPE_DELETE_OPERATION } from '../modules/deleteOperation/deleteOperation-types';
 import { buildEntityData } from './data-builder';
+import { elConvertHits, elConvertHitsToMap } from './engine-data-converter';
 import { internalEngineMappingGenerator } from './engine-mapping-generator';
 import {
   buildDataRestrictions,
@@ -1071,130 +1068,6 @@ export const RUNTIME_ATTRIBUTES = {
     `,
     getParams: async (context, user) => getRuntimeUsers(context, user),
   },
-};
-
-// region relation reconstruction
-const elBuildRelation = (type, connection) => {
-  return {
-    [type]: null,
-    [`${type}Id`]: connection.internal_id,
-    [`${type}Role`]: connection.role,
-    [`${type}Name`]: connection.name,
-    [`${type}Type`]: connection.types.find((connectionType) => !isAbstract(connectionType)),
-  };
-};
-const elMergeRelation = (concept, fromConnection, toConnection) => {
-  if (!fromConnection || !toConnection) {
-    throw DatabaseError('Reconstruction of the relation fail', concept.internal_id);
-  }
-  const from = elBuildRelation('from', fromConnection);
-  from.source_ref = `${convertTypeToStixType(from.fromType)}--temporary`;
-  const to = elBuildRelation('to', toConnection);
-  to.target_ref = `${convertTypeToStixType(to.toType)}--temporary`;
-  return R.mergeAll([concept, from, to]);
-};
-export const elRebuildRelation = (concept) => {
-  if (concept.base_type === BASE_TYPE_RELATION) {
-    const { connections } = concept;
-    const entityType = concept.entity_type;
-    const fromConnection = R.find((connection) => connection.role === `${entityType}_from`, connections);
-    const toConnection = R.find((connection) => connection.role === `${entityType}_to`, connections);
-    const relation = elMergeRelation(concept, fromConnection, toConnection);
-    relation.relationship_type = relation.entity_type;
-    return R.dissoc('connections', relation);
-  }
-  return concept;
-};
-const elDataConverter = (esHit, withoutRels = false) => {
-  const elementData = esHit._source;
-  const data = {
-    _index: esHit._index,
-    _id: esHit._id,
-    id: elementData.internal_id,
-    sort: esHit.sort,
-    ...elRebuildRelation(elementData),
-  };
-  const entries = Object.entries(data);
-  const ruleInferences = [];
-  for (let index = 0; index < entries.length; index += 1) {
-    const [key, val] = entries[index];
-    if (key.startsWith(RULE_PREFIX)) {
-      const rule = key.substring(RULE_PREFIX.length);
-      const ruleDefinitions = Object.values(val);
-      for (let rIndex = 0; rIndex < ruleDefinitions.length; rIndex += 1) {
-        const { inferred, explanation } = ruleDefinitions[rIndex];
-        const attributes = R.toPairs(inferred).map((s) => ({ field: R.head(s), value: String(R.last(s)) }));
-        ruleInferences.push({ rule, explanation, attributes });
-      }
-      data[key] = val;
-    } else if (key.startsWith(REL_INDEX_PREFIX)) {
-      // Rebuild rel to stix attributes
-      if (withoutRels) {
-        delete data[key];
-      } else {
-        const rel = key.substring(REL_INDEX_PREFIX.length);
-        const [relType] = rel.split('.');
-        data[relType] = isSingleRelationsRef(data.entity_type, relType) ? R.head(val) : [...(data[relType] ?? []), ...val];
-      }
-    } else {
-      data[key] = val;
-    }
-  }
-  if (ruleInferences.length > 0) {
-    data.x_opencti_inferences = ruleInferences;
-  }
-  if (data.event_data) {
-    data.event_data = JSON.stringify(data.event_data);
-  }
-  return data;
-};
-// endregion
-
-export const elConvertHitsToMap = async (elements, opts) => {
-  const { mapWithAllIds = false } = opts;
-  const convertedHitsMap = {};
-  let startProcessingTime = new Date().getTime();
-  for (let n = 0; n < elements.length; n += 1) {
-    const element = elements[n];
-    convertedHitsMap[element.internal_id] = element;
-    if (mapWithAllIds) {
-      // Add the standard id key
-      if (element.standard_id) {
-        convertedHitsMap[element.standard_id] = element;
-      }
-      // Add the stix ids keys
-      (element.x_opencti_stix_ids ?? []).forEach((id) => {
-        convertedHitsMap[id] = element;
-      });
-    }
-    // Prevent event loop locking more than MAX_EVENT_LOOP_PROCESSING_TIME
-    if (new Date().getTime() - startProcessingTime > MAX_EVENT_LOOP_PROCESSING_TIME) {
-      startProcessingTime = new Date().getTime();
-      await new Promise((resolve) => {
-        setImmediate(resolve);
-      });
-    }
-  }
-  return convertedHitsMap;
-};
-
-export const elConvertHits = async (data, opts = {}) => {
-  const { withoutRels = false } = opts;
-  const convertedHits = [];
-  let startProcessingTime = new Date().getTime();
-  for (let n = 0; n < data.length; n += 1) {
-    const hit = data[n];
-    const element = elDataConverter(hit, withoutRels);
-    convertedHits.push(element);
-    // Prevent event loop locking more than MAX_EVENT_LOOP_PROCESSING_TIME
-    if (new Date().getTime() - startProcessingTime > MAX_EVENT_LOOP_PROCESSING_TIME) {
-      startProcessingTime = new Date().getTime();
-      await new Promise((resolve) => {
-        setImmediate(resolve);
-      });
-    }
-  }
-  return convertedHits;
 };
 
 // elFindByIds is not defined to use ordering or sorting (ordering is forced by creation date)
